Moved simulator folder to components and simulator core to utils
