+++ src/components/simulator/types.ts
